[LCR-165] refactor: AuthService 인터페이스 도입 및 AuthServiceImpl 분리
